chore: move rule components
